Добавлена функция createValuesTable(), добавлена очистка данных датчика в таблице Values при удалении датчика, удаление всех данных в таблице Values при clearFromTables()

+++ SmartHomeConfig.cpp
@@ -1334,6 +1334,11 @@ bool SmartHomeConfig::deleteSensor(QTreeWidgetItem* item)
 
     if(sensor != nullptr)
     {
+        if(!deleteValueFromTable(sensor))
+        {
+            return false;
+        }
+
         if(!deleteSensorFromTable(sensor))
         {
             return false;
